cleanup
git-svn-id: https://only.mawhrin.net/repos/FBReaderJ/trunk@399 6a642e6f-84f6-412e-ac94-c4a38d5a04b0

+++ src/org/zlibrary/text/model/ZLTextModel.java
@@ -18,7 +18,7 @@ public interface ZLTextModel {
 	void addControl(ZLTextForcedControlEntry entry);
 	void addHyperlinkControl(byte textKind, String label);
 	void addImage(String id, Map<String,ZLImage> imageMap, short vOffset);
-	void addFixedHSpace(byte length);
+	void addFixedHSpace(short length);
 
 	String dump();
 }
